+++ Changelog.md
@@ -1,6 +1,6 @@
 # Unreleased / 2014-07-123
 
-* `preprocesser` option (mainly to use [`css-whitespace`](https://github.com/reworkcss/css-whitespace))
+* `transform` option (mainly to use [`css-whitespace`](https://github.com/reworkcss/css-whitespace))
 
 # 1.0.0 / 2014-07-14